feat: support x86-64 and arm64 architectures of macOS

+++ src/paimon/common/predicate/literal_converter_test.cpp
@@ -137,10 +137,12 @@ TEST_F(LiteralConverterTest, TestIntLiteral) {
 TEST_F(LiteralConverterTest, TestBigIntLiteral) {
     auto field_array =
         arrow::ipc::internal::json::ArrayFromJSON(arrow::int64(), R"([4, 5, null])").ValueOrDie();
-    CheckResult(field_array,
-                std::vector<Literal>({Literal(4l), Literal(5l), Literal(FieldType::BIGINT)}));
-    CheckLiteralsFromString(FieldType::BIGINT, {"4", "5"},
-                            std::vector<Literal>({Literal(4l), Literal(5l)}));
+    CheckResult(field_array, std::vector<Literal>({Literal(static_cast<int64_t>(4)),
+                                                   Literal(static_cast<int64_t>(5)),
+                                                   Literal(FieldType::BIGINT)}));
+    CheckLiteralsFromString(
+        FieldType::BIGINT, {"4", "5"},
+        std::vector<Literal>({Literal(static_cast<int64_t>(4)), Literal(static_cast<int64_t>(5))}));
     CheckLiteralFromRow(arrow::int64(),
                         {static_cast<int64_t>(4), static_cast<int64_t>(5), NullType()},
                         FieldType::BIGINT,
@@ -176,11 +178,11 @@ TEST_F(LiteralConverterTest, TestDoubleLiteral) {
     CheckLiteralsFromString(FieldType::DOUBLE, {"4.05", "5.17"},
                             std::vector<Literal>({Literal(4.05), Literal(5.17)}));
     CheckLiteralFromRow(arrow::float64(),
-                        {4.05, 5.17, static_cast<double> INFINITY, static_cast<double>(-INFINITY),
-                         static_cast<double> NAN, NullType()},
+                        {4.05, 5.17, static_cast<double>(INFINITY), static_cast<double>(-INFINITY),
+                         static_cast<double>(NAN), NullType()},
                         FieldType::DOUBLE,
-                        {Literal(4.05), Literal(5.17), Literal(static_cast<double> INFINITY),
-                         Literal(static_cast<double>(-INFINITY)), Literal(static_cast<double> NAN),
+                        {Literal(4.05), Literal(5.17), Literal(static_cast<double>(INFINITY)),
+                         Literal(static_cast<double>(-INFINITY)), Literal(static_cast<double>(NAN)),
                          Literal(FieldType::DOUBLE)});
 }
