+++ src/paimon/common/reader/predicate_batch_reader_test.cpp
@@ -96,40 +96,45 @@ TEST_F(PredicateBatchReaderTest, TestSimple) {
     {
         auto reader =
             std::make_unique<MockFileBatchReader>(data_array, data_type_, /*batch_size=*/10);
-        auto predicate = PredicateBuilder::LessThan(/*field_index=*/1, /*field_name=*/"f1",
-                                                    FieldType::BIGINT, Literal(24l));
+        auto predicate =
+            PredicateBuilder::LessThan(/*field_index=*/1, /*field_name=*/"f1", FieldType::BIGINT,
+                                       Literal(static_cast<int64_t>(24)));
         auto expected_array = std::make_shared<arrow::ChunkedArray>(data_array->Slice(0, 24));
         CheckResult(std::move(reader), predicate, expected_array);
     }
     {
         auto reader =
             std::make_unique<MockFileBatchReader>(data_array, data_type_, /*batch_size=*/10);
-        auto predicate = PredicateBuilder::LessThan(/*field_index=*/1, /*field_name=*/"f1",
-                                                    FieldType::BIGINT, Literal(1l));
+        auto predicate =
+            PredicateBuilder::LessThan(/*field_index=*/1, /*field_name=*/"f1", FieldType::BIGINT,
+                                       Literal(static_cast<int64_t>(1)));
         auto expected_array = std::make_shared<arrow::ChunkedArray>(data_array->Slice(0, 1));
         CheckResult(std::move(reader), predicate, expected_array);
     }
     {
         auto reader =
             std::make_unique<MockFileBatchReader>(data_array, data_type_, /*batch_size=*/10);
-        auto predicate = PredicateBuilder::LessThan(/*field_index=*/1, /*field_name=*/"f1",
-                                                    FieldType::BIGINT, Literal(99l));
+        auto predicate =
+            PredicateBuilder::LessThan(/*field_index=*/1, /*field_name=*/"f1", FieldType::BIGINT,
+                                       Literal(static_cast<int64_t>(99)));
         auto expected_array = std::make_shared<arrow::ChunkedArray>(data_array->Slice(0, 99));
         CheckResult(std::move(reader), predicate, expected_array);
     }
     {
         auto reader =
             std::make_unique<MockFileBatchReader>(data_array, data_type_, /*batch_size=*/10);
-        auto predicate = PredicateBuilder::GreaterThan(/*field_index=*/1, /*field_name=*/"f1",
-                                                       FieldType::BIGINT, Literal(0l));
+        auto predicate =
+            PredicateBuilder::GreaterThan(/*field_index=*/1, /*field_name=*/"f1", FieldType::BIGINT,
+                                          Literal(static_cast<int64_t>(0)));
         auto expected_array = std::make_shared<arrow::ChunkedArray>(data_array->Slice(1, 99));
         CheckResult(std::move(reader), predicate, expected_array);
     }
     {
         auto reader =
             std::make_unique<MockFileBatchReader>(data_array, data_type_, /*batch_size=*/10);
-        auto predicate = PredicateBuilder::GreaterThan(/*field_index=*/1, /*field_name=*/"f1",
-                                                       FieldType::BIGINT, Literal(98l));
+        auto predicate =
+            PredicateBuilder::GreaterThan(/*field_index=*/1, /*field_name=*/"f1", FieldType::BIGINT,
+                                          Literal(static_cast<int64_t>(98)));
         auto expected_array = std::make_shared<arrow::ChunkedArray>(data_array->Slice(99, 1));
         CheckResult(std::move(reader), predicate, expected_array);
     }
@@ -148,8 +153,9 @@ TEST_F(PredicateBatchReaderTest, TestVariousBatchSize) {
     ])")
                           .ValueOrDie();
     auto expected_array = std::make_shared<arrow::ChunkedArray>(PrepareArray(8));
-    auto predicate = PredicateBuilder::GreaterThan(/*field_index=*/1, /*field_name=*/"f1",
-                                                   FieldType::BIGINT, Literal(-1l));
+    auto predicate =
+        PredicateBuilder::GreaterThan(/*field_index=*/1, /*field_name=*/"f1", FieldType::BIGINT,
+                                      Literal(static_cast<int64_t>(-1)));
     for (auto batch_size : {5, 10, 11, 20}) {
         auto reader = std::make_unique<MockFileBatchReader>(data_array, data_type_, batch_size);
         CheckResult(std::move(reader), predicate, expected_array);
@@ -174,16 +180,18 @@ TEST_F(PredicateBatchReaderTest, TestFullAndEmptyCase) {
     {
         auto reader =
             std::make_unique<MockFileBatchReader>(data_array, data_type_, /*batch_size=*/10);
-        auto predicate = PredicateBuilder::LessThan(/*field_index=*/1, /*field_name=*/"f1",
-                                                    FieldType::BIGINT, Literal(20l));
+        auto predicate =
+            PredicateBuilder::LessThan(/*field_index=*/1, /*field_name=*/"f1", FieldType::BIGINT,
+                                       Literal(static_cast<int64_t>(20)));
         auto expected_array = std::make_shared<arrow::ChunkedArray>(data_array);
         CheckResult(std::move(reader), predicate, expected_array);
     }
     {
         auto reader =
             std::make_unique<MockFileBatchReader>(data_array, data_type_, /*batch_size=*/10);
-        auto predicate = PredicateBuilder::GreaterThan(/*field_index=*/1, /*field_name=*/"f1",
-                                                       FieldType::BIGINT, Literal(20l));
+        auto predicate =
+            PredicateBuilder::GreaterThan(/*field_index=*/1, /*field_name=*/"f1", FieldType::BIGINT,
+                                          Literal(static_cast<int64_t>(20)));
         CheckResult(std::move(reader), predicate, nullptr);
     }
 }
