+++ src/paimon/global_index/lumina/lumina_global_index_test.cpp
@@ -374,14 +374,15 @@ TEST_F(LuminaGlobalIndexTest, TestInvalidInputs) {
         {
             ASSERT_OK_AND_ASSIGN(auto reader,
                                  CreateGlobalIndexReader(index_root, data_type_, options_, meta));
-            ASSERT_NOK_WITH_MSG(reader->VisitVectorSearch(std::make_shared<VectorSearch>(
-                                    "f1",
-                                    /*limit=*/2, query_, /*filter=*/nullptr,
-                                    PredicateBuilder::Equal(/*field_index=*/1, /*field_name=*/"f0",
-                                                            FieldType::BIGINT, Literal(5l)),
-                                    /*distance_type=*/std::nullopt,
-                                    /*options=*/std::map<std::string, std::string>())),
-                                "lumina index not support predicate in VisitVectorSearch");
+            ASSERT_NOK_WITH_MSG(
+                reader->VisitVectorSearch(std::make_shared<VectorSearch>(
+                    "f1",
+                    /*limit=*/2, query_, /*filter=*/nullptr,
+                    PredicateBuilder::Equal(/*field_index=*/1, /*field_name=*/"f0",
+                                            FieldType::BIGINT, Literal(static_cast<int64_t>(5))),
+                    /*distance_type=*/std::nullopt,
+                    /*options=*/std::map<std::string, std::string>())),
+                "lumina index not support predicate in VisitVectorSearch");
         }
         {
             ASSERT_OK_AND_ASSIGN(auto reader,
